revert to double Atm

+++ src/main/java/interfaces/tes/Atm.java
@@ -13,7 +13,7 @@ public interface Atm<T extends State> {
 	
 	Optional<Integer[][]> getOccurrencesMatrix();
 	
-	BigDecimal[][] getMatrix();
+	Double[][] getMatrix();
 
-	BigDecimal[][] getMatrixCopy();
+	Double[][] getMatrixCopy();
 }
